Fix condition

+++ packages/basic-crawler/src/internals/basic-crawler.ts
@@ -835,8 +835,10 @@ export class BasicCrawler<Context extends CrawlingContext = BasicCrawlingContext
         // pass in failed requests back to the `crawler.run()`, otherwise they would be considered as handled and
         // ignored - as a failed requests is still handled.
         if (this.hasFinishedBefore && this.requestQueue?.name === 'default' && purgeRequestQueue) {
-            await this.requestQueue.drop();
-            this.requestQueue = await this._getRequestQueue();
+            if (this.requestQueue?.name === 'default' && purgeRequestQueue) {
+                await this.requestQueue.drop();
+                this.requestQueue = await this._getRequestQueue();
+            }
 
             this.stats.reset();
             await this.stats.resetStore();
